+++ Aplication/Lotes/RegistrarLoteQuery.cs
@@ -47,7 +47,7 @@ namespace Aplication.Lotes
 
         public long CultivoId { get; set; }
         public string Nombre { get; set; }
-        public int NumeroHectareas { get; set; }
+        public decimal NumeroHectareas { get; set; }
     }
     public class RegistrarLoteResponse
     {
