+++ Aplication/Palmas/EditarPalmaQuery.cs
@@ -33,7 +33,7 @@ namespace Aplication.Palmas
     public class EditarPalmaRequest : IRequest<EditarPalmaResponse>
     {
         public long PalmaId { get; set; }
-        public int Altura { get; set; }
+        public decimal Altura { get; set; }
         public string Descripcion { get; set; }
         public DateTime FechaSiembra { get; set; }
         public string Estado { get; set; }
