+++ Domain/Lotes/Lote.cs
@@ -10,7 +10,7 @@ namespace Domain.Lotes
 {
     public class Lote
     {
-        public Lote(long cultivoId, string nombre, int numeroHectareas)
+        public Lote(long cultivoId, string nombre, decimal numeroHectareas)
         {
             CultivoId = cultivoId;
             Nombre = nombre;
@@ -21,7 +21,7 @@ namespace Domain.Lotes
 
         public long Id { get; private set; }
         public string Nombre { get; private set; }
-        public int NumeroHectareas { get; private set; }
+        public decimal NumeroHectareas { get; private set; }
         public string Estado { get; private set; }
         public long CultivoId { get; private set; }
         public virtual Cultivo Cultivo { get; private set; }
@@ -29,7 +29,7 @@ namespace Domain.Lotes
         private readonly List<Palma> _palmas;
         public IReadOnlyCollection<Palma> Palmas => _palmas.AsReadOnly();
 
-        public void Editar(string nombre, int numeroHectareas, string estado)
+        public void Editar(string nombre, decimal numeroHectareas, string estado)
         {
             Nombre = nombre;
             NumeroHectareas = numeroHectareas;
